core/permissions.py
from rest_framework.permissions import BasePermission


class IsCitizen(BasePermission):
    def has_permission(self, request, view):
        return request.user and request.user.is_authenticated and request.user.role == "citizen"


class IsOfficer(BasePermission):
    def has_permission(self, request, view):
        return request.user and request.user.is_authenticated and request.user.role in {"officer", "admin"}


class IsOwnerOrOfficer(BasePermission):
    def has_permission(self, request, view):
        return request.user and request.user.is_authenticated

    def has_object_permission(self, request, view, obj):
        if request.user and request.user.is_authenticated and request.user.role in {"officer", "admin"}:
            return True
        return obj.citizen_id == request.user.id
