requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
djangorestframework-simplejwt>=5.3,<6.0
drf-spectacular>=0.27,<1.0
whitenoise>=6.6,<7.0
PyMySQL>=1.1.0,<2.0
python-dotenv>=1.0,<2.0
